+++ packages/vfs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tinycloud/vfs
 
+## 0.1.13-beta.9
+
+### Patch Changes
+
+- Updated dependencies [d894c57]
+  - @tinycloud/node-sdk@2.11.0-beta.11
+
 ## 0.1.13-beta.8
 
 ### Patch Changes